C API: (lec)        - use a std::deque<double> object instead of a double* array          for the execution stack        - rewrite all executable LEC elements (functions, operators,          time functions and multi-arg functions) for clarity and to          use the new stack implementation
fix #1187

+++ api/lec/l_exec_val.cpp
@@ -1,26 +1,3 @@
-/**
- * @header4iode
- *
- * Functions to evaluate LEC "values". 
- * 
- * The table (*L_VAL_FN)[] contains the function pointers to evaluate L_PI to L_EURO (see iode.h).
- *
- * Function signatures:
- *
- *      double  <fnname>(int t);
- *  or
- *      double  <fnname>(void);
- *  
- *  List of functions
- *  -----------------
- *      double L_pi ()
- *      double L_euro()
- *      double L_e ()
- *      double L_time(int t)
- *      double L_i(int t)
- *     
- */
-
 #include "api/lec/lec.h"
 
 
@@ -28,17 +5,17 @@
 
 double L_pi(int t) 
 { 
-    return (double) M_PI; 
+    return M_PI; 
 }
 
 double L_euro(int t) 
 { 
-    return (double) 40.3399; 
+    return 40.3399; 
 }
 
 double L_e(int t) 
 { 
-    return (double) M_E; 
+    return M_E; 
 }
 
 double L_time(int t)
